+++ instrumentation/net/http/otelhttp/internal/semconv/dup.go
@@ -15,6 +15,7 @@
 package semconv // import "go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp/internal/semconv"
 
 import (
+	"io"
 	"net/http"
 	"strings"
 
@@ -164,11 +165,19 @@ func (d dupHTTPServer) TraceResponse(resp ResponseTelemetry) []attribute.KeyValu
 			semconvNew.HTTPRequestBodySize(resp.ReadBytes),
 		)
 	}
+	if resp.ReadError != nil && resp.ReadError != io.EOF {
+		// This is not in the semantic conventions, but is historically provided
+		attributes = append(attributes, attribute.String("http.read_error", resp.ReadError.Error()))
+	}
 	if resp.WriteBytes > 0 {
 		attributes = append(attributes,
 			semconvOld.HTTPResponseContentLength(resp.WriteBytes),
 			semconvNew.HTTPResponseBodySize(resp.WriteBytes),
 		)
+	}
+	if resp.WriteError != nil && resp.WriteError != io.EOF {
+		// This is not in the semantic conventions, but is historically provided
+		attributes = append(attributes, attribute.String("http.write_error", resp.WriteError.Error()))
 	}
 	if resp.StatusCode > 0 {
 		attributes = append(attributes,